# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API сервиса автоматического управления ценами на маркетплейсах Wildberries и Ozon.\n\n## Аутентификация\nПосле успешного входа (POST /api/auth/login) сервер устанавливает HttpOnly-cookie `rx_session`.\nБраузер отправляет его автоматически; при работе через curl/Postman передавайте `-b \"rx_session=\u003ctoken\u003e\"`.\n\n## CSRF-защита\nВсе мутирующие эндпоинты (POST/PATCH/DELETE), кроме `/api/auth/register`, `/api/auth/login`\nи `/api/auth/verification/resend`, проверяют заголовок `Origin`.\nОн должен совпадать с одним из разрешённых источников (настраивается через `ALLOWED_ORIGINS`).\nSwagger UI выставляет `Origin` автоматически.",
        "title": "RepricerX API",
        "contact": {
            "name": "Akim Zuev",
            "email": "akim.zuev.86@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/shops": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Возвращает все магазины, принадлежащие аутентифицированному пользователю.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shops"
                ],
                "summary": "Список магазинов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/transport.shopResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Сессия отсутствует или истекла",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Создаёт новый магазин, сохраняет API-ключи в зашифрованном виде (AES-GCM).\n\nФормат поля `credentials` зависит от маркетплейса:\n- **wb** (Wildberries): `{\"token\": \"\u003cAPI-токен\u003e\"}`\nТокен можно получить в личном кабинете WB: Настройки → Доступ к API.\n- **ozon**: `{\"client_id\": \"\u003cClient-ID\u003e\", \"api_key\": \"\u003cAPI-ключ\u003e\"}`\nКлючи доступны в кабинете Ozon Seller: Настройки → API-ключи.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shops"
                ],
                "summary": "Подключить магазин",
                "parameters": [
                    {
                        "description": "Данные нового магазина",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.createShopRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Магазин создан",
                        "schema": {
                            "$ref": "#/definitions/transport.shopResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат или неизвестный маркетплейс (код: bad_request / invalid_marketplace)",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Не аутентифицирован",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "403": {
                        "description": "CSRF — Origin не совпадает",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/shops/{id}": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Возвращает магазин по UUID. Доступен только владельцу.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shops"
                ],
                "summary": "Получить магазин",
                "parameters": [
                    {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000",
                        "description": "UUID магазина",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transport.shopResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат UUID",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Не аутентифицирован",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Магазин не найден (код: shop_not_found)",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Удаляет магазин и все связанные данные.\nЕсли за последние 30 дней есть записи в журнале изменений цен — операция будет отклонена.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shops"
                ],
                "summary": "Удалить магазин",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID магазина",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Магазин удалён"
                    },
                    "400": {
                        "description": "Неверный формат UUID",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Не аутентифицирован",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "403": {
                        "description": "CSRF — Origin не совпадает",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Магазин не найден",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Изменяет имя, учётные данные и/или настройки автообновления и расписания.\nВсе поля опциональны — передавайте только то, что хотите изменить.\n\nПоле `scheduleCron` задаёт расписание пересчёта цен в формате cron (5 полей).\nПримеры: `\"0 * * * *\"` — каждый час, `\"0 */4 * * *\"` — каждые 4 часа.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shops"
                ],
                "summary": "Обновить магазин",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID магазина",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновляемые поля",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.updateShopRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transport.shopResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Не аутентифицирован",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "403": {
                        "description": "CSRF или нет доступа к магазину",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Магазин не найден",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/shops/{id}/test": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Выполняет тестовый запрос к API маркетплейса с сохранёнными учётными данными.\nПри успехе статус магазина становится `active`, при ошибке авторизации — `error`.\nРезультат проверки фиксируется в поле `lastCheckedAt`.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shops"
                ],
                "summary": "Проверить подключение к маркетплейсу",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID магазина",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Подключение успешно",
                        "schema": {
                            "$ref": "#/definitions/transport.testConnectionResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат UUID",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Не аутентифицирован",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "403": {
                        "description": "CSRF — Origin не совпадает",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Магазин не найден",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "422": {
                        "description": "Ошибка авторизации в маркетплейсе (код: auth_failed)",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "transport.createShopRequest": {
            "type": "object",
            "required": [
                "credentials",
                "marketplace",
                "name"
            ],
            "properties": {
                "credentials": {
                    "type": "object"
                },
                "marketplace": {
                    "type": "string",
                    "example": "wb"
                },
                "name": {
                    "type": "string",
                    "example": "Мой магазин на WB"
                }
            }
        },
        "transport.errorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "invalid_credentials"
                },
                "message": {
                    "type": "string",
                    "example": "Неверный email или пароль"
                }
            }
        },
        "transport.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/transport.errorDetail"
                }
            }
        },
...
        "transport.shopResponse": {
            "type": "object",
            "properties": {
                "autoUpdateEnabled": {
                    "type": "boolean",
                    "example": false
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "lastCheckedAt": {
                    "type": "string"
                },
                "marketplace": {
                    "type": "string",
                    "example": "wb"
                },
                "name": {
                    "type": "string",
                    "example": "Мой магазин на WB"
                },
                "scheduleCron": {
                    "type": "string",
                    "example": "0 */4 * * *"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
        "transport.testConnectionResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
...
        "transport.updateShopRequest": {
            "type": "object",
            "properties": {
                "autoUpdateEnabled": {
                    "type": "boolean",
                    "example": true
                },
                "credentials": {
                    "type": "object"
                },
                "name": {
                    "type": "string",
                    "example": "Новое название"
                },
                "scheduleCron": {
                    "type": "string",
                    "example": "0 */4 * * *"
                }
            }
        }
    },
    "securityDefinitions": {
        "SessionCookie": {
            "description": "Сессионный cookie, выставляемый при логине (POST /api/auth/login).",
            "type": "apiKey",
            "name": "rx_session",
            "in": "cookie"
        }
    }
}